Modularize the wetty service and add events (#18)
* - Add help message
- Removed unnecesary ifs

*     Moved to a module.
    Add events:
    * exit (code): Ssh process finished, informs the process exitcode.
    * disconnect : Socket disconnected.

* Fix on previous commit

* Fix on previous commit
